==> 1_frame_trimming.Rmd <==
```{r}
suppressMessages(suppressWarnings(library(tidyverse)))
```

```{r}
# read and trim tracking data
weeks <- seq(1:8)
for(w in weeks){
    
  # read each week  
  df_tracking <- read_csv(paste0("week",w,".csv"),
                                col_types = cols())
  
  # flip positional values such that offense always faces right
  df_tracking <- df_tracking %>%
    mutate(x = ifelse(playDirection == "left", 120 - x, x),
           y = ifelse(playDirection == "left", 160 / 3 - y, y))
  
  df_tracking <- df_tracking %>%
    mutate(dir = ifelse(playDirection == "left", dir + 180, dir),
           dir = ifelse(dir > 360, dir - 360, dir),
           o = ifelse(playDirection == "left", o + 180, o),
           o = ifelse(o > 360, o - 360, o))
  
  # identify frames by snap and end event
  passer_tracking_data <- df_tracking %>%
    mutate(is_start = as.numeric(event %in% c("autoevent_ballsnap", "ball_snap")),
           is_end = as.numeric(event %in% c("fumble", "handoff", "lateral",
                                            "autoevent_passforward", "pass_forward",
                                            "qb_sack", "qb_strip_sack", "run")))
  
  # summarize each play by starting and ending frame ID
  frame_seq_info <- passer_tracking_data %>% 
    group_by(gameId, playId) %>%
    mutate(any_start = any(is_start == 1), any_end = any(is_end == 1)) %>%
    filter(any_start, any_end) %>%
    summarize(start_frame = frameId[which(is_start == 1)[1]],
              end_frame = frameId[which(is_end == 1 & frameId > start_frame)[1]], 
              .groups = "drop")
  
  # find the starting and ending frame id for tracking data
  passer_tracking_seq_data <- passer_tracking_data %>%
    left_join(frame_seq_info, by = c("gameId", "playId")) %>%
    filter(!is.na(start_frame), !is.na(end_frame),
           frameId >= start_frame, frameId <= end_frame)
  
  # reindex frame ids such that first frame has ID = 1
  passer_tracking_seq_data <- passer_tracking_seq_data %>%
    mutate(new_frameId = frameId - start_frame + 1)
 
   # drop the extra rows
  cut_by_frame_clean <- passer_tracking_seq_data %>% 
  select(-c("is_start", "is_end", "start_frame", "end_frame", "frameId"))
  
  # rename column to frameID
  colnames(cut_by_frame_clean)[colnames(cut_by_frame_clean) == "new_frameId"] = "frameId"
  
  # create csv
  write.csv(cut_by_frame_clean, paste0("week",w,"_cut_by_frame_clean.csv"), 
            row.names = FALSE)
}
```
